+++ app/components/badges-page/badge-card.hbs
@@ -5,8 +5,10 @@
   ...attributes
 >
   {{#if this.userHasBadgeAward}}
+    {{! @glint-expect-error: not ts-ified yet }}
     <BadgesPage::EarnedBadgePreview @badge={{@badge}} class="mb-3 lg:mb-6" />
   {{else}}
+    {{! @glint-expect-error: not ts-ified yet }}
     <BadgesPage::LockedBadgePreview @badge={{@badge}} class="mb-3 lg:mb-6" />
   {{/if}}
